requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-environ==0.4.54
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.13
